this won't work, gotta change directions
I'm gonna use the std lambda functions, trying to get lambdas working
without them is taking way too much time.

I'm hoping to implement my own libraries, for the sake of learning, but
implementing lambdas with the ability to bind to a `this` context just
isn't making sense to me.