some fixes for lapse lib
* lapse::array::operator+= wasn't working at all, guess I hadn't tested
it yet. Seems good now.

* added f32_obj.is_inf(), f32_obj.is_finite(), 32_obj.is_nan()

* u32_to_c_str() wasn't working with trailing zeroes (1'000'000 return
"1"), works now

* added some tests to lapse test for u32_to_c_str(). Really need to get
string comparison up so these tests can be automated, at the moment it's
up to you to look at them and confirm the strings match the numbers

* str::operator+ doesn't return a reference anymore, was having lifetime
issues with that func, need to think it through some more sometime

* fixed a cyclical dependency between math and f32 by moving the relevant
method declarations into the cpp files and only including the headers of
the other files in the cpp files. (the math logarithm func is using
f32_obj to test for floating poitn infinity at the moment)