giving up on opengl
This commit doesn't work. It's just a blank glfw window. Should be
rendering a

can't get opengl textures to render for mysterious reasons. All I want
is a plot() function, so, opengl should've just been a simple way to get
that.

* Added stb_image lib
* Added code that initializes opengl, loads in a texture, and then tries
to display that texture (but fails)
* Added test_image.png (it's a public domain image from project
gutenberg)

After this I'll revert all opengl code and bring in something else. I'm
just committting this so I could theoretically return to it someday.