Added minesweeper stuff
* Added sooth_image
* Added a platform::draw_bitmap that takes an image and screen position
for drawing it
* I drew some simple bitmap images for minesweeper. They ought to do I
think, though they're not much good
* Drawing a grid of images in the window
* Added minesweeper project

I'm implementing minesweeper inside of sooth_engine for the moment, I'll
move it into the minesweeper project and turn sooth_engine into a static
lib once it's working properly. I think this will be simpler than making
it a library while I implement it.