Playable minesweeeper
*Added sooth mouth is_mouse_left_button_hit() &c funcs. Need to think
about how this should be structured. Having seprate functions for each
button is a little silly.
*Moved sooth mouse out of sooth input (specifying Input::Mouse:: each
time was redundant)
*Added minesweeper::tile_obj struct for tracking tile information. Using
just an enum for it proved inconvenient
*Added right & left clicking on minesweeper tiles
*Added chain reactions of tile reveals when you click on an empty tile
*Started moving code into minesweeper namespace
*Altered the minesweeper png background colors to show which ones are
revealed (blue background) and which ones are hidden (desaturated
background)

Minesweeper doesn't end the game, doesn't reveal all bombs when you hit
a bomb, but otherwise it's pretty much playable.