fix mouse hovering on 0th row & column
off by 1 error. Also switched to using signed integers since that's what
we're receiving from the platform anyway. Currently it's being converted
from an i32 to an f32 then back to an i32 as it travels through the
layers of abstration between the platform and the actual game ... so
that's something.