minesweeper now resets properly for each game
So when you hit a bomb it generates a new map. It also closes and
reopens the game window, which maybe it shouldn't be doing that.

*Added game_session and game_run ideas. A session starts at application
startup, a run starts at the start of gameplay.
*So initializing the game session is when you load in files and setup
the program, then the game run initialization is used for
setting/resettign game mechanic varaibles
*added reset() function to tile_obj for reseting it to the default
values, had to copy/paste them from the struct, should think about
better ways to do this
*fixed a minor spelling error rode -> rcode

Had to put the grid variable declaration in a weird place to resolve a
C++ declaration order cyclical dependency issue. Should be solvable by
moving things into separate files.

Most of this diff is copy/pasting code from one area of the codebase to
another.