Add 1d6 safe spaces to first click
* Added tile_obj.to_rect()
* Added array.sample()
* Added template to lapse::rect. Unfortunately this caused a few diffs
for adding the <> template to other files. Ideally this would've been
it's own commit to clean things up a bit.
* Some unrelated in progress changes to mui_params where I'm trying to
get the box model working ... This change had to be included becuase've
the rect<> change above.

This uses a simple "growing into adjacent spaces" algorithm for adding
1d6 safe spaces to the starting area.

The 3x3 starting area didn't work well when combined with maps where
that safe area didn't trigger any reveal chain reactions.

I'm not sure if "frontier" is the most understandable term for the
"spaces adjacent to the current set of safe spaces" idea.

I've heard of this algorithm being called Pan's Algorithm when it's used
for generating mazes, but I've also heard that term used for an
unrelated algorithm for generating mazes so I'm never sure what the name
of this particular algorithm is.