add i32_min, other changes in support of mui dev
* array.last_index() now returns an i32 instead of a u32. I think I want
to go through and remove most of the unsigned ints I've sprinkled in
thoughout the codebase. There are probably a few that make sense (like
for array lengths) but most of them were a bad idea, it just causes
warnings in the compiler and potential bugs, but there isn't really any
upside to using them.
* add f32 version of max. Should really go through a performance pass on
the lapse math library someday
* add i32_min. Lined up whitespace on integer constants.