mui div auto sizing to fit contents
* Added mui::draw() this allows us to defer drawing mui elements until
later, which allows us to calculate the size of a div before drawing it
* Not really happy with the current mui::draw implementation, a
mui::DrawCall is something like 120 bytes for some reason (I think the
lambda is the problem). In principle the lambda could be replaced with a
simple enum and switch case, so, it's not a big issue. I am curious why
the C++ std lambda would take up so many bytes.
* Moved the code for is_hovering() is_active() into mui context (was in
button.cpp before)

The mui code is currently very messy. Things like the box_property &
mui_size structs probably don't make sense, I wrote a lot of this while
I was very confused.

There are some safe spaces changes in this commit too:

* Added debug key: press the 1 key to toggle highlighting the safe spaces
(I need to debug why the safe spaces frontier is able to expand
diagonally ... it's still doing that sometimes)
* Cleaned up the safe spaces generation code a bit ... these changes are
"dark work," they don't actually alter the outcome of the function, but
previous the safe_3x3 variable was markig spaces outside of the game map
as "safe," and they were just being ignored because of the call to
.adjacents(), but now it limits the 3x3 to the game board.
* Increased the size of the safe space, from 1d6 to 6d5