mui divs auto-size, with padding & border
- mui supports having multiple containers (multiple divs)
- Things are being properly positioned with respect to padding & border
now.
- I think margin is more-or-less working now
- The box_property struct was returning vertical properties as horizontal
& vice versa. Reversed it to fix them. Hadn't used the individual
properties before so I hadn't noticed.
- mui_button now uses two elements: one for the button rect and one for
the text
- Text now fits tightly into its element. There's about 1 pixel of space
on top and bottom.
- mui::draw_text() supports font_color

I'm not really happy with the way mui is shaping up. The params struct
is enormous and constantly growing. Need to break things up a lot more,
maybe have a stack of mui properties instead, and just pop those off for
the draw calls and just have them line up with the proper draw calls (so
each draw could would be popping off the correct number of properties it
needs, and that number would happen to be on top when the draw call is
made, etc.

The sizeof() the mui structs is silly already, and they don't have half
the stuff they would in a complete implementation of this idea.