scale pixel size to fit window onto screen
currently we just assume the user's monitor is 1080p, since that's
likely. Need to see if pixel engine can get us the actual information.

the "huge game" size is just a window that's 80% of a 1080p display.
Might change it to be something like 1080p minus 20 pixels of padding or
something like that. Might want more padding than that though.

Would be neat to see a map with pixel scales below 1.0, but I'm not sure
if that works. Accidentally did it with a map 16x bigger than the
screen, but at that size each tile was only one pixel so you couldn't
make out anything of the game state.

The font color in the pngs needs to be changed to support the huge map
size, cause the numbers become unreadable on larger maps, not
distinguished enough from their background.