tell the player when they win
victory.png comes from project gutenberg's THE VICTORY AT SEA by REAR-ADMIRAL W. S. SIMS U.S. NAVY (which is in the public domain). Specifically it's a snippet of a picture of the front cover of the book (the word victory).

also added a "quick game" option to the new game menu, which makes a 5x5
game. I lowered the safe space size back down to 1d6 to avoid an
infinite loop on quick games. Need to have it scale the size of the safe
space to suit the map size.