colorful menu, better organization of menu code
- planing on adding a color_scheme.h file with preprocessor definitions
  of the colors in the color scheme... Not sure what color scheme to use
  currently.

The program is currently crashing pretty much all the time. Not sure
exactly why, but I've known for awhile that I have more new commands
than delete commands I'm running, I've just been ignoring that issue,
cause a lot of it will go away when I add in arena allocation stuff and
temporary allocation stuff, but this seems to be a different issue
related to how the olc app closes and reopens when you start a new game.

I know the olc pixel engine uses threading ... and I know nothing about
threading ... so it could be something to do with that.