add & to params : removing unneccessary copies
- Added constants for the menu strings
  - Wasn't able to declare them const because C++ is silly, but they won't
change so I don't know that it matters.
  - I'm not sure if these constants should be declared in the .h or .cpp. I
did some in one place and some in the other. I tend to think the .cpp
makes more sense, but maybe I should have them as a separate struct in
the .h file or something
- Added an assert if sooth image fails to load an image
- removed the numbers from the grid tile enum
- added an assert on array.reserve() to call attention to any reserve(0)
calls being made, since that shouldn't really be happening
- There's a bunch of debugging code in this commit, logging to the
console when constructors are called and whatnot. I think I'll remove
that before merging the branch, but not before
- I'm not sure I understand the idea of copy constructors, are they
supposed to make a copy by value or by reference? I went with by value
for now, but I'm not sure if that's "right."

Haven't worked on arrays yet. This is just strings and images.

Still don't really deallocate memory when the program is closed.