faster fill_rect
memcpy() & hoisting variables out of the loop

- Had to add some conditionals to fill_rect cause with the faster code I
got crashes when it went over the edges. Not sure why that wasn't an
issue before
- Added impossible game mode (2x 1080p, can't play it myself since I have
a 1080p monitor, also I think it doesn't render correctly so it's a
little unplayable at the moment).
- Lowered temp arena size. I was wrong, it was only using like 50 bytes
before actually
- Commented out std::cout calls in the string class, those calls were the
lion's share of the performance cost of string calls at the moment so it
was disrupting the msvc performance stuff