faster draw_line
from ~197 ms to ~47 ms, very roughly. I don't think the msvc performance
measurement thing is that accurate when dealing with numbers so low. But
these sorts of changes made the other funcs faster so they should help
here.

One reason why I'm dubious about the msvc numbers: draw_rect (which just
calls draw_line four times) is exactly the same time as draw_line (so...
calling the function four times takes as many milliseconds as calling it
once?)

Function just isn't called that many times to do too much performance
testing on it though. I think it'll just be used in UIs though so it's not
pressing that be improved a lot.