Off by one errors in ui
- Removed -1s from lapse rect, that was the wrong solution
- Added -1 to draw_rect size. Because if you send in a position of 0, 0
and a size of 5, 5, you expect a 5px by 5px square ... but you would've
gotten a 6px by 6px square before
- Had an issue with fill_rect floats memcpy()ing fractions of pixels.
Switched to ints.
- Added margin-bottom: 10px to menu buttons (except the last menu
button).

It's hard for me to understand these off by one errors, not sure how
many of them there are right now in the ui, most of it looks okay, but
I'm not entirely sure.

The scale of the UI is problematic for these errors as well, whenever
the scale results in a resolution that's not evenly divisible by the
number of elements being drawn in the UI, they have to become oddly
mis-sized, with padding: 10px becoming 8px in some places and 9px in
other places.