working on off by one issue in ui
I'm now rounding the numbers after scaling, but rounding down the size
if the size exceeds the screen size. I could just -1 the size if it
passes the screen size ...

and even this isn't a real solution, because if either width or height
exceeds screen size I round both down, when in reality you need to round
down only the one that exceeds the screen size, otherwise you risk an
off by one error in how the things are drawn ... but it's only at the
edges of the screen, so, I dunno, maybe it's fine to just have a broken
edge case there. I could obviously add more branches to the code here to
account for the two edge cases, but I would need to construct a testing
apparatus to confirm that it's working properly.

I think this is probably fine for now