Well that code doesn't do anything after all
Code from my last commit was wrong, and it never enters that branch
anyway.

I noticed I was using ceil() when I should've used floor() and it made
me realize this code wasn't actually going off. I'm confused by that,
cause the game used to crash because it was hitting this situation. I
guess the rounding thing solves it better than I thought.

It needs further investigation with clearer test cases.

Added todo about it, but it's not a high priority.

-renamed Console::log() to print()