Began work on draw_bitmap_rotated()
-I tried taking the current draw_bitmap_scaled() and add rotation to it,
but I found it difficult to understand how to modify that function in
its current state, so I went back to an earlier version of draw bitmap,
and made draw_bitmap_rotated, which is based off of the old unoptimized
draw_bitmap. It doesn't do any rotation yet, just draws the sprite.
-I reorganized draw_bitmap_scaled a bit, trying to clean up the code.
-Added images, a ground tile, bullet, etc
-I increased the temp arena to 10 million bytes again, because
draw_bitmap_rotated() is using it for the moment.