draw_bitmap_rotated() now rotates in screen space
-added a yellow border along the bullet sprite's edge to highlight the
errors in the previous rotation code
-had to add a new git ignore for some visual studio temp folder thing

now there is a 'dithering' sortof glitch with the rotation where it
isn't drawing all the pixels in the new image location (because there
are more pixels in the rotated image than there were in the unrotated
image). It's particularly visible if you shoot the bullet at 45 degree
angles (because a diagonal movement is longer than a cardinal movement on a
square grid). At least I think that's what's going on.