draw_bitmap_rotated(): not avging missing pixels
instead of averaging a "missing pixel" between the two adjacent pixels,
we're now merging those two pixels based off of how close the missing
pixel's floating point location is to them each. (so if it's 90% close
to one then that pixel gets 90% influence over the color of the missing
pixel).

Visual inspection shows this change not working, but conceptually it
makes sense, so I think the issue is something else. I think I might be
merging the pixels in the wrong order, since logically each of the image
pixels might be transparent, in which case you'd need to blend each of
them with the screen pixels before blending the two pixels together to
find the missing pixel ... if that's the case, then the solution might
be to simply make a list of the missing pixels on the first pass, that
way all the image pixels will be drawn on the screen, and then on a
second pass draw the missing pixels.