e&g: enemy bullets damage player
The bullet collision rects are way off because they are coming from the
sprites, but the bullet sprites are one tile in size, but that is mostly
transparency, the bullets themselves are much smaller than a full sprite
tile.

Currently this is drawing the bullet collision rects to highly the
issue. There's another issue in that the bullet collision rects are not
rotating with the bullets (which could be fixed by making enemy
projectiles round or something).